Fix the original incorrect implemenet by twos complement
The original assumption is that A and B hold one integer,
and then they turn that integer to two's complement as the vector in the original paper.

This still works. However, when these input vectors add with the random vectors,
elements in the result vector must exceed the range of the binary number:

  Xa + Ra = Xa'  ---> {3} + {9}  = {12}, pass to B
  Xb + Rb = Xb'  ---> {1} + {13} = {14}, pass to A

It doesn't help if we change the number to binary and do vector addition:

  Xa + Ra = Xa' ---> {0 0 1 1} + {1 0 0 1} = {1 0 1 2}
  Xb + Rb = Xb' ---> {0 0 0 1} + {1 1 0 1} = {1 1 0 2}

And it makes it worse: since Xa' and Xb' are passed to its counterpart,
the counterpart can easily know the secret input bit is 1 if in the result vector it is 2.

This is exactly the security issue that in the original paper:

	Wenliang Du and Zhijun Zhan. 2002.

	A practical approach to solve Secure Multi-party Computation problems.

	In Proceedings of the 2002 workshop on New security paradigms (NSPW '02).
	Association for Computing Machinery, New York, NY, USA, 127–135.
	https://doi.org/10.1145/844102.844125

In 4.1:

	"if the elements of Y are in the domain of [0,100], and we also know the random numbers are generated from
	[0, 200] domain, then if an element of vector Y + Rb is 250, we know the original element in vector Y is bigger than 50"

The key here is the domain of elements of the random vector and the input vector must be large enough to project such a leak.
Turning the input and random vector to binary numbers, and use them as the vectors to add, is actually the worst case.

----

However, if we only turn them as binary numbers, but do following the normal binary addition when adding two vectors,
we won't reduce the strength of security, and we can gain a little performance improvement by the fact that:

    Dot product of two binary vectors = bitwise AND, and then bitcount

For example:

    Xa + Ra = {{0 0 1 1}} + {{1 0 0 1}} = {{1 1 0 0}}, do binary number addition, and then pass to B as Xa'

    Xa' . Xb = {{1 1 0 0}} . {{0 0 0 1}} = {{0}} = {{0 0 0 0}}, because Xa' and Xb have no bit 1 at the same position.

Binary vectors' dot product is like:

    {{1 1 0 0 }} . {{1 0 0 0}} = (1 * 1) + (1 * 0) + (0 * 0) + (0 * 0) = how many bits with 1 at the same position

Because dot product is:

    a1 * b1 + a2 * b2 + .... an * bn

When ai and bi are binary, only when both positions' bits are 1, the value counts.
Otherwise, the value will be 0. So dot producting two binary vectors is equal to count how many
bit 1 at the same position of both vectors. This operation can be done by:

    1. bitwise AND both binary vectors
    2. Do bitcount like the trick at: http://graphics.stanford.edu/~seander/bithacks.html#CountBitsSetParallel

That is, if we turn every element in SMC scalar-product protocol into binary vector:

    Xa                                    + Ra
    {{1 0 0 1} {0 0 1 1}} .... {1 1 0 1}} + {{1 0 1 1} {0 0 1 0} .... {1 0 0 0}}
    {9         3          .... 13       } + {11        2         .... 8        }

Although we still need to follow the binary number addition to prevent information leak:

    Xa'
    {{1 0 1 0 0} {0 0 1 0 1} .... {{1 0 1 0 1}}
    {20          5           .... 21          }

The following step `Xa' . Xb` can be boosted a little by the fact that binary dot product = AND + bitcount.
However, this really makes it faster "a little" when the two elements are integer.
Because integer multiplication is highly optimized nowadays.
Maybe float elements can be boosted significantly by turning them to binary representations.